+++ src/main.rs
@@ -1,34 +1,16 @@
+mod args;
 mod fmt_mount;
 mod json;
 
 use {
-    argh::FromArgs,
-    std::{cmp::Reverse, fs, os::unix::fs::MetadataExt, path::PathBuf},
+    crate::args::*,
+    crossterm::tty::IsTty,
+    std::{
+        cmp::Reverse,
+        fs,
+        os::unix::fs::MetadataExt,
+    },
 };
-
-#[derive(FromArgs)]
-/// List your filesystems.
-///
-/// All units are SI.
-///
-/// Source at https://github.com/Canop/lfs
-struct Args {
-    /// print the version
-    #[argh(switch, short = 'v')]
-    version: bool,
-
-    /// whether to show all mount points
-    #[argh(switch, short = 'a')]
-    all: bool,
-
-    /// output as JSON
-    #[argh(switch, short = 'j')]
-    json: bool,
-
-    #[argh(positional)]
-    /// if a path is provided, only the device holding this path will be shown
-    pub path: Option<PathBuf>,
-}
 
 fn main() -> lfs_core::Result<()> {
     let args: Args = argh::from_env();
@@ -59,6 +41,8 @@ fn main() -> lfs_core::Result<()> {
         Ok(())
     } else {
         mounts.sort_by_key(|m| Reverse(m.size()));
-        fmt_mount::print(&mounts)
+        let color = args.color.value()
+            .unwrap_or_else(|| std::io::stdout().is_tty());
+        fmt_mount::print(&mounts, color)
     }
 }